+++ src/domain/challenge/challenge/icon/ChallengeIcon.tsx
