+++ src/domain/challenge/hub/icon/HubIcon.tsx
