+++ src/domain/challenge/opportunity/icon/OpportunityIcon.tsx
