+++ src/domain/collaboration/canvas/icon/CanvasIcon.tsx
