+++ src/domain/platform/admin/toplevel/constants.ts
@@ -2,7 +2,7 @@ import { TabDefinition } from '../../../../common/components/core/PageTabs/PageT
 import PeopleOutlinedIcon from '@mui/icons-material/PeopleOutlined';
 import ForumOutlinedIcon from '@mui/icons-material/ForumOutlined';
 import GppGoodOutlinedIcon from '@mui/icons-material/GppGoodOutlined';
-import { HubIcon } from '../../../shared/icons/HubIcon';
+import { HubIcon } from '../../../challenge/hub/icon/HubIcon';
 
 export enum AdminSection {
   Hub = 'hubs',
