+++ src/domain/platform/breadcrumbs/useBreadcrumbs.ts
@@ -6,8 +6,8 @@ import {
   useHubNameQuery,
   useOpportunityNameQuery,
 } from '../../../core/apollo/generated/apollo-hooks';
-import { HubIcon } from '../../shared/icons/HubIcon';
-import { ChallengeIcon } from '../../shared/icons/ChallengeIcon';
+import { HubIcon } from '../../challenge/hub/icon/HubIcon';
+import { ChallengeIcon } from '../../challenge/challenge/icon/ChallengeIcon';
 import { EntityTypeName } from '../../shared/layout/PageLayout/SimplePageLayout';
 
 export interface BreadcrumbsItem {
