+++ src/domain/shared/components/search-cards/SearchChallengeCard.tsx
@@ -1,6 +1,6 @@
 import React, { FC } from 'react';
 import HubOutlinedIcon from '@mui/icons-material/HubOutlined';
-import { ChallengeIcon } from '../../icons/ChallengeIcon';
+import { ChallengeIcon } from '../../../challenge/challenge/icon/ChallengeIcon';
 import SearchJourneyWithParentCard, { SearchJourneyWithParentImplProps } from './SearchJourneyWithParentCard';
 
 export interface SearchChallengeCardProps extends SearchJourneyWithParentImplProps {}
